Refactor: 게시글 상세 및 작성 페이지 컴포넌트를 분리하여 구조 개선 및 메타데이터 추가 #41
